move IPython.zmq.entry_point to IPython.utils.kernel
